requirements.txt
pandas
lxml
pypdf
streamlit
fastapi
uvicorn
sentence-transformers
faiss-cpu
python-dotenv
openai
tqdm

pymupdf
pytesseract
pillow